requirements-dev.txt
googleapis-common-protos>=1.6.0
matplotlib>=3.1.1
numpy>=1.16.4
opentracing>=2.2.0
opentelemetry-api>=1.10.0
opentelemetry-sdk>=1.10.0
opentelemetry-exporter-otlp-proto-grpc>=1.10.0
protobuf>=3.8.0
psutil>=5.6.3
pytest>=5.0.1
requests>=2.22.0
urllib3>=1.25.3
flake8>=3.7.8
lightstep>=4.2.0
